🧪 [Tests]: #125 [FE] DetailPanel labels テストを labels.test.tsx に分離+重複追加ケース
rendering.test.tsx に同居していた labels 追加/削除のテストを
DetailPanel.labels.test.tsx に移設。境界値として「既存ラベルと同一文字列を
追加しても onTaskUpdate が呼ばれない」を新規追加。

Refs #125
